requirements.txt
pandas==2.3.3
python-dotenv==1.2.1
python-telegram-bot==22.6
starlette>=0.40,<1.0
uvicorn>=0.30,<1.0
